+++ src/main/java/org/smoothbuild/task/exec/ArtifactBuilder.java
@@ -39,6 +39,7 @@ public class ArtifactBuilder {
   }
 
   public void runBuild() {
+    artifactSaver.clean();
     taskBatch.executeAll();
     if (!taskBatch.containsErrors()) {
       console.println("built artifact(s):");