+++ src/main/java/org/smoothbuild/task/save/ArtifactSaver.java
@@ -4,6 +4,7 @@ import static java.util.stream.Collectors.joining;
 import static org.smoothbuild.io.fs.base.Path.path;
 import static org.smoothbuild.task.save.ArtifactPaths.artifactPath;
 import static org.smoothbuild.task.save.ArtifactPaths.targetPath;
+import static org.smoothbuild.task.save.ArtifactPaths.toFileName;
 
 import java.util.Set;
 
@@ -13,6 +14,7 @@ import org.smoothbuild.cli.Console;
 import org.smoothbuild.io.fs.base.FileSystem;
 import org.smoothbuild.io.fs.base.Path;
 import org.smoothbuild.lang.function.base.Name;
+import org.smoothbuild.lang.type.ArrayType;
 import org.smoothbuild.lang.type.Type;
 import org.smoothbuild.lang.type.TypesDb;
 import org.smoothbuild.lang.value.Array;
@@ -33,34 +35,39 @@ public class ArtifactSaver {
     this.console = console;
   }
 
+  public void clean() {
+    fileSystem.delete(artifactPath(Path.root()));
+  }
+
   public void save(Name name, Value value) {
+    Path path = path(toFileName(name));
     if (value instanceof Array) {
-      saveArray(name, (Array) value);
+      saveArray(path, (Array) value);
     } else if (value.type().equals(typesDb.file())) {
-      saveValue(name, ((Struct) value).get("content"));
+      saveBasicValue(path, ((Struct) value).get("content"));
     } else {
-      saveValue(name, value);
+      saveBasicValue(path, value);
     }
   }
 
-  private void saveArray(Name name, Array array) {
+  private void saveArray(Path path, Array array) {
     Type elemType = array.type().elemType();
-    if (elemType.equals(typesDb.file())) {
-      saveFileArray(name, array);
+    fileSystem.createDir(artifactPath(path));
+    if (elemType instanceof ArrayType) {
+      int i = 0;
+      for (Array element : array.asIterable(Array.class)) {
+        saveArray(path.append(path(Integer.toString(i))), element);
+        i++;
+      }
+    } else if (elemType.equals(typesDb.file())) {
+      saveFileArray(path, array);
     } else {
-      saveValueArray(name, array);
+      saveValueArray(path, array);
     }
   }
 
-  private void saveValueArray(Name name, Array array) {
-    Path artifactPath = artifactPath(name);
-
-    fileSystem.delete(artifactPath);
-
-    // Create dir explicitly. When fileArray is empty for loop below won't
-    // create empty dir for us.
-    fileSystem.createDir(artifactPath);
-
+  private void saveValueArray(Path path, Array array) {
+    Path artifactPath = artifactPath(path);
     int i = 0;
     for (Value value : array.asIterable(Value.class)) {
       Path filePath = path(Integer.valueOf(i).toString());
@@ -71,16 +78,9 @@ public class ArtifactSaver {
     }
   }
 
-  private void saveFileArray(Name name, Array fileArray) {
+  private void saveFileArray(Path path, Array fileArray) {
     DuplicatesDetector<String> duplicatesDetector = new DuplicatesDetector<>();
-
-    Path artifactPath = artifactPath(name);
-    fileSystem.delete(artifactPath);
-
-    // Create dir explicitly. When fileArray is empty for loop below won't
-    // create empty dir for us.
-    fileSystem.createDir(artifactPath);
-
+    Path artifactPath = artifactPath(path);
     for (Struct file : fileArray.asIterable(Struct.class)) {
       Path sourcePath = artifactPath.append(path(((SString) file.get("path")).data()));
       if (!duplicatesDetector.addValue(((SString) file.get("path")).data())) {
@@ -91,19 +91,18 @@ public class ArtifactSaver {
 
     if (duplicatesDetector.hasDuplicates()) {
       Set<String> duplicates = duplicatesDetector.getDuplicateValues();
-      console.error(duplicatedPathsMessage(name, duplicates));
+      console.error(duplicatedPathsMessage(path, duplicates));
     }
   }
 
-  private String duplicatedPathsMessage(Name name, Set<String> duplicates) {
+  private String duplicatedPathsMessage(Path path, Set<String> duplicates) {
     String separator = "\n  ";
     String list = separator + duplicates.stream().collect(joining(separator));
-    return "Can't store result of '" + name + "' as it contains files with duplicated paths:"
-        + list;
+    return "Can't store array of Files as it contains files with duplicated paths:" + list;
   }
 
-  private void saveValue(Name name, Value value) {
-    Path artifactPath = artifactPath(name);
+  private void saveBasicValue(Path path, Value value) {
+    Path artifactPath = artifactPath(path);
     Path targetPath = targetPath(value);
     fileSystem.delete(artifactPath);
     fileSystem.createLink(artifactPath, targetPath);